Output TrackId of spatially nested parent
In cases where a subhalo is nested in another subhalo but HBT has
not detected this from the merger history, we now output the
TrackId of the possible new parent as NewParentTrackId.

The subhalo hierarchy is not modified in this version and the
radius used is just the half mass radius from the last snapshot.