Modify search for subhalos to re-nest
For each subhalo we now try to find the least massive, more
massive subhalo which encloses this one. Subhalo A encloses B
if B's centre is within twice the half mass radius of A.

This done in ascending order of subhalo mass: the least massive
subhalos are checked for re-nesting first.