Disable subsampling for certain particle types. (#86)
* Use std::partition instead of using own implementation.

Previous implementation differs in the ordering of bound particles within Elist relative to new implementation. I believe this is because it first started from the end of Elist.

* Move particle shuffling to its own class.

* Move BHs to beginning of Particle vector and shuffle the rest.

* Formatting.

* change variable name

* Scale masses of just the subsampled set of particles.

* Increase tree size to contain MaxSampleSize + Nunsample particles

This change also handles the assignment of particle_mass in the next loop.

* Make removal of unbound particles keep original relative ordering.

* Update value of NewNunsample after an unbinding iteration

* Fixed potential bug

I was looping over Particles directly to count number of bound particles that we cannot subsample. However, this method could've include unbound particles in the count.

* Only subsample if we have more particles than Nsubsample + Nunsample

* Fix loop limits when computing Msubsample.

* New runtime input to specify particle types that cannot be subsampled.

* Code now uses bit mask.

* Allow the user to disable new behaviour by using -1 value.

* Silly ~ instead of ! to turn a true to a false.

...
Cannot believe I spent the last hour doing this.

* Add sanity checks for runtime parameters.

This commit also restructures how parameters are tested.

* Check for duplicate values in vector input parameters

* Additional check on SnapshotIdList

* Remove Swift parameter

* Enable the user to declare NO_STRIPPING within CMAKE.